refactor(deposit): merge providers
